Add generic JupyterLab introduction / tutorial
This adds a basic introduction or tutorial for JupyterLab.
It gives an overview of how to use the environment, what extra options
exists (such as ssh access), what some of the terminology used is, and
what to watch out for.

Detailed tutorials on specific (analysis) uses should be provided as
separate documents.