Initial setup for Kubernetes in the HPCCloud
This provides a simple, initial set up for putting Kubernetes, with a
JupyterHub on top of it, on the SURFSara HPC cloud, including an
extensive description(*).

The setup includes Python scripts to start and shut down the virtual
machines (using the OpenNebula API), shell scripts to execute the
Kubernetes and Helm commands (such as installing Helm and JupyterHub),
and an Ansible playbook to set up the controller and worker nodes.

(*) This description is very rough, and should be read as a write-down
of steps tried and followed, and problems encountered (and solved or
worked-around) along the way.