Provide a stylesheet for the documentation
The stylesheet is mainly to be used by ASCIIDoctor when generating the
help information.

The stylesheet is a rather horrible hack: it works from the default
JupyterHub stylesheet (since it also used for the JupyterHub login
page), but adds a few things from the EUCP website. The latter uses a
Bootstrap theme, but items added to this stylesheet are very minimal.

Ideally, the same stylesheet(s) as for the EUCP website should be
used, but these feel horrendously bloated for simple text-based
documentation; hence there are only a few style items used instead.