Handle linking to README or index in subdir
This handles proper linking to the index file in subdirectories. We
can't set outfilesuffix to '/', since this doesn't work for every link
(some links point to files instead of directories), thus we create a
new attribute 'indexfile', that is set to index, but to README for the
proper (GitHub) environments.